pyproject.toml
[tool.black]
exclude = '\.venv/|\.eggs/|\.env/'

[tool.isort]
multi_line_output = 3
include_trailing_comma = true

[tool.poetry]
name = "StrainMap"
version = "1.2.3"
description = ""
homepage = "https://github.com/imperialcollegelondon/strainmap"
authors = ["RSE Team, Research Computing Service, Imperial College London <ict-rse-team@imperial.ac.uk>"]
license = "GPL-3.0-or-later"
readme = "README.md"
repository = "https://github.com/ImperialCollegeLondon/strainmap"
classifiers = [
	"Development Status :: 5 - Production/Stable",
    "Environment :: MacOS X",
    "Environment :: Win32 (MS Windows)",
    "Environment :: X11 Applications",
	"Intended Audience :: End Users/Desktop",
	"Intended Audience :: Science/Research"
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
scipy = "^1.9.1"
xarray = "^2022.6.0"
tensorlayer = "<=2.2"
tensorflow = "<=2.8"
matplotlib = "^3.5.3"
pydicom = "^2.3.0"
nibabel = "^4.0.2"
netCDF4 = "^1.6.1"
natsort = "^8.2.0"
opencv-python = "^4.6.0"
openpyxl = "^3.0.10"
python-decouple = "^3.6"
tqdm = "^4.64.1"
keras = "<=2.8"
h5py = "<=3.6"
protobuf = "<=3.20"
PyPubSub = "^4.0.3"

[tool.poetry.dev-dependencies]
black = "*"
flake8 = "<5.0.0"
mypy = "*"
pyinstaller = "*"
pytest = "*"
pytest-cov = "*"
pytest-flake8 = "*"
pytest-mock = "*"
pytest-mypy = "*"
pytest-runner = "*"
pre-commit = "^2.20.0"
isort = "^5.10.1"
bump2version = "^1.0.1"
mkdocs = "^1.3.1"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}
mkdocs-material = "^8.5.0"
mkdocs-gen-files = "^0.4.0"
mkdocs-literate-nav = "^0.4.1"
mkdocs-section-index = "^0.3.4"

[tool.poetry.scripts]
strainmap = "strainmap.entrypoints:strainmap"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"